requirements.txt
streamlit 
selenium
groq
beautifulsoup4
lxml 
html5lib
python-dotenv